+++ src/masternodes/rpc_customtx.cpp
@@ -206,7 +206,7 @@ public:
 
     void operator()(const CGovernanceHeightMessage& obj) const {
         rpcInfo.pushKV(obj.govVar->GetName(), obj.govVar->Export());
-        rpcInfo.pushKV("startHeight", obj.startHeight);
+        rpcInfo.pushKV("startHeight", static_cast<uint64_t>(obj.startHeight));
     }
 
     void operator()(const CAppointOracleMessage& obj) const {
